[Terrain] heightmap displacement in VS, restructure code
