+++ Data/EngineSettings.ini
@@ -14,8 +14,8 @@ EnvironmentMapResolution=512
 Width=1600
 Height=900
 DisplayMode=Windowed
-PreferredDisplay=0
-Scene=0
+PreferredDisplay=1
+Scene=4
 
 DebugWindow=false
 DebugWindowWidth=450
