+++ Data/Scenes.ini
@@ -3,3 +3,4 @@ Default=0
 Sponza=1
 EnvironmentMapUnitTest=2
 StressTest=3
+Terrain=4
