+++ Source/Renderer/Renderer.h
@@ -119,7 +119,10 @@ enum EBuiltinPSOs // TODO: remove the hardcoded PSOs when a generic Shader solut
 	DOWNSAMPLE_DEPTH_CS_PSO,
 	DEBUGVERTEX_LOCALSPACEVECTORS_PSO,
 	DEBUGVERTEX_LOCALSPACEVECTORS_PSO_MSAA_4,
-
+	TERRAIN,
+	TERRAIN_MSAA4,
+	TERRAIN_TESSELLATED,
+	TERRAIN_TERRAIN_TESSELLATED_MSAA4,
 	NUM_BUILTIN_PSOs
 };
 
@@ -144,6 +147,7 @@ enum EBuiltinRootSignatures
 	LEGACY__FFX_FSR1,
 	LEGACY__UI_HDR_Composite,
 	LEGACY__DownsampleDepthCS,
+	LEGACY__Terrain,
 
 	NUM_BUILTIN_ROOT_SIGNATURES
 };
